+++ src/components/ContentPopup.tsx
@@ -69,7 +69,7 @@ function ContentPopup({ trigger, setTrigger, bgColor = 'bg-dark-sea', children }
             <button
               className="z-[9999] modest-shadow absolute top-4 right-4 w-10 h-10 p-2 bg-bright-seaweed rounded-full transition-colors hover:xs:bg-hovered-seaweed"
               onClick={() => setTrigger(false)}>
-              <CloseSvg width="100%" height="auto" fill="#fff" />
+              <CloseSvg width="100%" height="100%" fill="#fff" />
             </button>
             {children}
           </motion.div>