+++ src/pages/api/auth/by-token.ts
@@ -13,7 +13,7 @@ type UserDataServerResponseProps = {
   first_name: string;
   last_name: string;
   email: string;
-  profile_picture: string | null;
+  cloud_pic: string | null;
 };
 
 type UserDataServerResponse = AxiosResponse<UserDataServerResponseProps>;