+++ src/pages/api/auth/refresh-token.ts
@@ -32,27 +32,33 @@ export default async function handler(_req: RefreshTokenRequest, res: RefreshTok
     return res.status(405).json({ errors: { message: 'Given request method is not allowed here.' } });
   }
 
-  return await axios
-    .post('https://cherrytomaten.herokuapp.com/authentication/token/refresh/', {
-      refresh: _req.body.refreshToken.token,
-    })
-    .then((_res: RefreshTokenServerResponse) => {
-      const accessToken = tokenDecompiler(_res.data.access);
-      const refreshToken = tokenDecompiler(_res.data.refresh);
+  try {
+    const refreshToken: TokenPayload = JSON.parse(_req.body.refreshToken);
 
-      if (accessToken === null || refreshToken === null) {
-        return res.status(500).json({ errors: { message: 'A server error occured.' } });
-      }
-      return res.status(_res.status).json({
-        userId: _req.body.refreshToken.userId,
-        access: accessToken,
-        refresh: refreshToken,
+    return await axios
+      .post('https://cherrytomaten.herokuapp.com/authentication/token/refresh/', {
+        refresh: refreshToken.token,
+      })
+      .then((_res: RefreshTokenServerResponse) => {
+        const accessToken = tokenDecompiler(_res.data.access);
+        const refreshToken = tokenDecompiler(_res.data.refresh);
+
+        if (accessToken === null || refreshToken === null) {
+          return res.status(500).json({ errors: { message: 'A server error occured.' } });
+        }
+        return res.status(_res.status).json({
+          userId: _req.body.refreshToken.userId,
+          access: accessToken,
+          refresh: refreshToken,
+        });
+      })
+      .catch((err: BackendErrorResponse) => {
+        if (err.response?.data?.detail === undefined) {
+          return res.status(err.response?.status ? err.response?.status : 500).json({ errors: { message: 'A server error occured.' } });
+        }
+        return res.status(err.response.status).json({ errors: { message: err.response.data.detail } });
       });
-    })
-    .catch((err: BackendErrorResponse) => {
-      if (err.response?.data?.detail === undefined) {
-        return res.status(err.response?.status ? err.response?.status : 500).json({ errors: { message: 'A server error occured.' } });
-      }
-      return res.status(err.response.status).json({ errors: { message: err.response.data.detail } });
-    });
+  } catch (_err) {
+    return res.status(400).json({ errors: { message: 'Wrong token format.' } });
+  }
 }