+++ src/pages/api/favorites/get.ts
@@ -3,31 +3,22 @@ import axios, { AxiosResponse } from 'axios';
 import { FetchServerErrorResponse } from '@/types/Server/FetchServerErrorResponse';
 import { AUTH_TOKEN } from "@/types/constants";
 import { TokenPayload } from "@/types/Auth/TokenPayloadProps";
+import { FavoritePinsList } from "@/types/Pins/FavoritePinsList";
 
-type AddFavoriteRequest = NextApiRequest & {
-  body: { pinId: string };
-};
-
-type AddFavoriteResponseBody = {
-  uuid: string;
-  pin: string;
-  user: string;
-};
-
-type AddFavoriteErrorResponse = {
+type GetFavoriteErrorResponse = {
   response: AxiosResponse<{ res: string }>;
 };
 
-type AddFavoriteResponse = NextApiResponse<AddFavoriteResponseBody | FetchServerErrorResponse>;
+type GetFavoriteResponse = NextApiResponse<FavoritePinsList | FetchServerErrorResponse>;
 
 /**
- * Add a mappoint to the current users favorite list
- * @param _req containing the pinId in the body
- * @param res returns ids of the involded user / pin to confirm the success and the uuid of the database entry for the favorite pin entry or an error response.
+ * Get all favorite mappoints from a user
+ * @param _req containing an access token in the header
+ * @param res List of all favorites or an error response
  */
-export default async function handler(_req: AddFavoriteRequest, res: AddFavoriteResponse) {
+export default async function handler(_req: NextApiRequest, res: GetFavoriteResponse) {
   // wrong request method
-  if (_req.method !== 'POST') {
+  if (_req.method !== 'GET') {
     return res.status(405).json({ errors: { message: 'Given request method is not allowed here.' } });
   }
 
@@ -40,17 +31,15 @@ export default async function handler(_req: AddFavoriteRequest, res: AddFavorite
     const authToken: TokenPayload = JSON.parse(authTokenString);
 
     return await axios
-      .post('https://cherrytomaten.herokuapp.com/api/favorites/pin/', {
-        pin: _req.body.pinId,
-      }, {
+      .get("https://cherrytomaten.herokuapp.com/api/favorites/pin/", {
         headers: {
-          Authorization: 'Bearer ' + authToken.token,
+          Authorization: "Bearer " + authToken.token,
         },
       })
-      .then((_res: AxiosResponse<AddFavoriteResponseBody>) => {
+      .then((_res: AxiosResponse<FavoritePinsList>) => {
         return res.status(201).json(_res.data);
       })
-      .catch((err: AddFavoriteErrorResponse) => {
+      .catch((err: GetFavoriteErrorResponse) => {
         if (err.response?.data?.res === undefined) {
           return res.status(err.response.status).json({ errors: { message: 'A server error occured.' } });
         }
@@ -60,14 +49,3 @@ export default async function handler(_req: AddFavoriteRequest, res: AddFavorite
     return res.status(400).json({ errors: { message: 'Wrong token format.' } });
   }
 }
-
-/*
-err:
-   server crash
-err duplicate:
-    res: string
-success:
-  "uuid": "055b232c-70c0-45e0-88f1-286b01dfc4fc",
-  "pin": "16170359-5830-412a-b0ec-dc7d2831723f",
-  "user": "0e5ed6b9-d5b7-47f0-a4fa-202973061946"
- */