+++ src/pages/api/pins/get.ts
@@ -43,7 +43,6 @@ export default async function handler(_req: PinsApiRequest, res: NextApiResponse
         },
       })
       .then((_res: any) => {
-        logger.log('Mappoints response data: ', _res.data);
         let pins: PinProps[] = _res.data;
 
         const pinsInRange: PinProps[] = pins.filter((pin) => {