+++ src/types/Pins/index.ts
@@ -1,8 +1,15 @@
-export type PinProps = MapPointProps & {
+export type PinProps = {
+  uuid: string;
+  title: string;
+  description: string;
+  address: string;
+  openingHours: OpeningProps;
+  ratings: RatingProps[];
+  comments: CommentProps[];
+  image: ImageProps[];
   category: string | null;
   creator: string;
   longitude: number;
   latitude: number;
-  notes: string | null;
   created: string;
 };